reorganize custom resolvers code
